move to RuleType.Serializer
